Move ZoneState to web models
