+++ src/main/java/com/rackspace/salus/monitor_management/web/model/ZoneUpdate.java
@@ -15,7 +15,6 @@
  */
 package com.rackspace.salus.monitor_management.web.model;
 
-import com.rackspace.salus.monitor_management.types.ZoneState;
 import com.rackspace.salus.monitor_management.web.model.validator.ValidCidrList;
 import java.util.List;
 import lombok.Data;
